Rename memcpy to local_memcpy in rs_allocations.c.
This is so it doesn't get linked with instead of the bionic version.
This fixes a crash when compiling with -O0, where the memcpy version
in rs_allocations.c was hanging around (not being inlined) and was
linked with in unrelated code, which led to a crash because the ABI
was not the expected.

Change-Id: I0c699bf3e0ed7b353d0e2108c52fcb19660eecc5
Signed-off-by: Verena Beckham <verena@codeplay.com>
(cherry picked from commit 6fb5a4f5411c16d36f123c7c2c9add4d937479ef)

Bug: 30312793